server/requirements.txt
# ====================================
# 🔧 Core Django & REST Framework
# ====================================
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# ====================================
# ☁️ Google Cloud Storage Integration
# ====================================
google-cloud-storage>=2.14.0
google-auth>=2.0.0
django-storages>=1.13.2

# ====================================
# 🎨 Django Admin Customization
# ====================================
django-jazzmin>=2.6.0

# ====================================
# 🔐 Environment Variables
# ====================================
python-dotenv>=1.0.0

# ====================================
# 🔑 JWT Authentication (Optional - if used)
# ====================================
djangorestframework-simplejwt>=5.3.0

# ====================================
# 🧪 Optional Testing Tools (Uncomment if needed)
# ====================================
# pytest
# pytest-django
